+++ core/lib/core.dart
@@ -1,0 +1,1 @@
+library core;
