+++ data/lib/data.dart
@@ -1,0 +1,1 @@
+library data;
