+++ domain/lib/domain.dart
@@ -1,0 +1,1 @@
+library domain;
