+++ navigation/lib/navigation.dart
@@ -1,0 +1,1 @@
+library navigation;
